requirements.txt
python-dotenv
pymystem3
pandas
gensim
streamlit